migrates to laravel 5.8

+++ PROJECTS/laravel/database/migrations/2017_02_14_222731_create_photos_table.php
@@ -14,15 +14,15 @@ class CreatePhotosTable extends Migration
     public function up()
     {
         Schema::create('photos', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->timestamps();
             $table->softDeletes();
             $table->string('name',128);
             $table->text('description')->nullable();
-            $table->integer('album_id')->unsigned();
+            $table->bigInteger('album_id')->unsigned();
             $table->string('img_path',128);
             $table->foreign('album_id')->on('albums')->references('id')->onDelete('cascade');
-            
+
         });
     }
 
@@ -34,7 +34,7 @@ class CreatePhotosTable extends Migration
     public function down()
     {
         \DB::statement('SET FOREIGN_KEY_CHECKS=0;');
-     
+
         Schema::dropIfExists('photos');
     }
 }
