+++ PROJECTS/laravel/database/migrations/2017_05_07_224936_create_models_album_categories_table.php
@@ -14,9 +14,9 @@ class CreateModelsAlbumCategoriesTable extends Migration
     public function up()
     {
         Schema::create('album_categories', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('category_name', 64)->unique();
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->on('users')->references('id');
             $table->softDeletes();
             $table->timestamps();
